solar_ecommerce/requirements.txt
Django>=5.1,<7.0
python-dotenv>=1.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
drf-spectacular>=0.27
Pillow>=10.0
whitenoise>=6.8
requests>=2.32,<3.0

# Production server
gunicorn>=22.0
psycopg2-binary>=2.9

# Testing
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
